% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{doEDA}
\alias{doEDA}
\title{Generate graphs for EDA with a line of code}
\usage{
doEDA(type, .tbl_df, ...)
}
\arguments{
\item{type}{a string for chart type. Options availble are \itemize{
\item{\strong{"pie"}} for pie chart
\item{\strong{"waffle"}} for waffle chart
\item{\strong{"hbar"}} for horizontal bar chart
\item{\strong{"tree"}} for tree map
}}

\item{.tbl_df}{data frame}

\item{...}{arguments passed down to functions \code{pieChart()},
\code{waffleChart()}, \code{hbarChart()} and \code{treeMap()}}
}
\value{
pie chart / horizontal bar chart via ggplot2 waffle chart via
  waffle::waffle treemap via treemap::treemap
}
\description{
Generate graphs for EDA with a line of code
}
\examples{
library(dplyr)
data("HairEyeColor")
df <- as.data.frame(HairEyeColor)
#pie chart
doEDA(type = "pie", .tbl_df = df, .group = Eye)

#waffle chart
doEDA("waffle", df, .group = Eye, divisor = 0.1)

#horizontal bar chart
doEDA(type = "hbar", .tbl_df = df, .innerGroup = vars(Hair, Eye, Sex),
.outerGroup = vars(Hair, Eye), .xVar = Sex, .fill = Hair)

#treemap
doEDA(type = "tree", .tbl_df = df, .group = Hair, index = 'Hair')
}
\seealso{
\link[ggplot2]{ggplot2}, \link[waffle]{waffle}, \link[treemap]{treemap}
}
