% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{hbarChart}
\alias{hbarChart}
\title{Visualize proportions of subgroup samples}
\usage{
hbarChart(.tbl_df, .innerGroup, .outerGroup, .xVar, .fill)
}
\arguments{
\item{.tbl_df}{A data frame}

\item{.innerGroup}{A list of variable names wrapped with
\code{\link[dplyr:vars]{vars()}}}

\item{.outerGroup}{A list of variable names wrapped with
\code{\link[dplyr:vars]{vars()}}}

\item{.xVar}{One of the variables from .innerGroup excluding the .fill
variable}

\item{.fill}{One of the variables from .innerGroup}
}
\value{
A \strong{ggplot} object
}
\description{
\code{hbarChart()} frees user from repetitive workflows for
  creating horizontal bar chart in order to visualize subgroup distributions.
}
\details{
.innerGroup should, as the argument suggests, be a list of variables
  for grouping. They are the unit groups. On the other hand, .outerGroup is a
  list of variables shorter than .innerGroup. In turn .outerGroup is at an
  outer level and .innerGroup is a subset of .outerGroup. All the variables
  passed into .innerGroup and .outerGroup must be selected by vars() as they
  are quosures inside the function body. You could modify the labels for y
  axis using \code{scale_y_continuous()} since it is a ggplot object.
}
\examples{
library(dplyr)
hbarChart(invest, vars(gender, age, investment, invest_plan, perc),
vars(investment, invest_plan), age, perc) +
ggplot2::facet_wrap(vars(invest_plan)) +
ggplot2::scale_fill_brewer(type = "qual", palette = "Blues")
}
