% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\docType{data}
\name{invest}
\alias{invest}
\title{Data on Investment preferences}
\format{A data frame with 14311 rows and 5 variables:
\describe{
\item{gender}{Female or Male}
\item{age}{5 segemented age range between 18 to 51+ years old}
\item{invest_plan}{future investment choices for financial products}
\item{perc}{percentage weightages on invest_plan}
\item{investment}{current investments}
}}
\usage{
data(invest)
}
\description{
A dummy dataset containing a genders, age, investment and other attributes on
investors.
}
\author{
广州金十信息科技有限公司 \email{job@jin10.com}
}
\keyword{data}
