% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek At Data As A Whole}
\usage{
peek(.tbl_df)
}
\arguments{
\item{.tbl_df}{data.frame object}
}
\value{
Return a ggplot object via \code{vis_dat} from \code{visdat}
}
\description{
Peek At Data As A Whole
}
\examples{
d <- data.frame(int = 1:10,
                string = letters[1:10],
                factor = factor(month.abb[1:10]),
                stringsAsFactors = FALSE)
peek(d)
}
