% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{pieChart}
\alias{pieChart}
\title{Draw a fast pie chart}
\usage{
pieChart(.tbl_df, .group, start = 0, direction = 1, labelon = TRUE,
  ...)
}
\arguments{
\item{.tbl_df}{A data frame}

\item{.group}{A \emph{factor} variable}

\item{start}{\code{\link[ggplot2:coord_polar]{coord_polar()}} argument to
control where the counting position starts. Defaults at 0}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{labelon}{logical scalar. Should labels be added via
\code{\link[ggplot2:geom_text]{geom_text()}}?}

\item{...}{arguments passed down to \code{\link[ggplot2:labs]{labs()}} for
\emph{title}, \emph{subtitle}, etc.}
}
\value{
A \strong{ggplot} object
}
\description{
\code{pieChart()} saves the repeating data manipulation and
  ggplot2 graphic grammars to produce a pie chart. It is for visualizing
  distribution of factor levels of a single variable.
}
\examples{
pieChart(invest, gender)
}
