% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chn.R
\name{theme_chn}
\alias{theme_chn}
\title{ggplot theme wrapper for display CJK characters}
\usage{
theme_chn(family = "STXihei")
}
\arguments{
\item{family}{font family type}
}
\value{
ggplot2 theme list
}
\description{
ggplot theme wrapper for display CJK characters
}
\examples{
NULL
}
