% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{treeMap}
\alias{treeMap}
\title{Fast tree map wrapper}
\usage{
treeMap(.tbl_df, .group, index, type = "index", sizeby = "n",
  fontfamily.labels, fontfamily.title, title = NULL, ...)
}
\arguments{
\item{.tbl_df}{A data frame}

\item{.group}{A factor variable}

\item{index}{A string identical to .group}

\item{type}{A string for index type}

\item{sizeby}{A string for which the tile size depending on}

\item{fontfamily.labels}{A string for font type}

\item{fontfamily.title}{A string for font type}

\item{title}{A title}

\item{...}{arguments passed down to \code{treemap()}}
}
\value{
A \strong{ggplot} object
}
\description{
This is a wrapper for the workflow producing a tree map via
  \code{treemap} package. See
  \code{\link[treemap:treemap-package]{treemap-package}}.
}
\details{
Other than most of the arguments inherited from treemap(), treeMap()
  embbeded a default font type, \emph{Helvetica} for Latin characters and
  \emph{STXihei} CJK characters on \strong{MacOS}, when
  \code{fontfamily.title} and \code{fontfamily.labels} are \strong{both} not
  specified. One should note that both \emph{fontfamily.*} arguments should
  be be absent one-sided. This will not change the font type according to
  user's will other than the default ones.
}
\examples{
treeMap(invest, investment, "investment")
}
\seealso{
\link[treemap:treemap]{treemap}
}
