% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{hbarChart}
\alias{hbarChart}
\title{Visualize proportions of subgroup samples}
\usage{
hbarChart(.tbl_df, ..., columns, .xVar, .fill)
}
\arguments{
\item{.tbl_df}{A data frame}

\item{...}{fixed columns names supplied from \code{.tbl_df}}

\item{columns}{numeric vector that has a length shorter than the supplied by
\code{...}}

\item{.xVar}{One of the variables from ... besides the .fill variable}

\item{.fill}{One of the variables from ...}
}
\value{
A \strong{ggplot} object
}
\description{
\code{hbarChart()} frees user from repetitive workflows for
  creating horizontal bar chart in order to visualize subgroup distributions.
}
\details{
.innerGroup should, as the argument suggests, be a list of variables
  for grouping. They are the unit groups. On the other hand, .outerGroup is a
  list of variables shorter than .innerGroup. In turn .outerGroup is at an
  outer level and .innerGroup is a subset of .outerGroup. All the variables
  passed into .innerGroup and .outerGroup must be selected by \code{vars()}
  as they are quosures inside the function body. You could modify the labels
  for y axis using \code{scale_y_continuous()} since it is a ggplot object.
}
\examples{
library(dplyr)
library(ggplot2)

hbarChart(invest, gender, age, investment, invest_plan, perc,
columns = 3:4, .xVar = age, .fill = perc) +
facet_wrap(vars(invest_plan)) +
scale_fill_brewer(type = "qual", palette = "Blues")
}
