% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_excel.R
\name{tidy_excel}
\alias{tidy_excel}
\title{Import Tidy Excel Sheets into R}
\usage{
tidy_excel(file, ...)
}
\arguments{
\item{file}{\strong{.xls} or \strong{.xlsx} file}

\item{...}{arguments passed down to \link[readxl]{read_excel()}}
}
\value{
A \link[tibble]{tibble}
}
\description{
Import Tidy Excel Sheets into R
}
\details{
This function wraps the steps for importing excel files having more
  than 1 sheet through \code{mutate()} a \emph{sheetName} column after
  intermediate iterations of sheet names via \code{map()}. For excel files
  that have a single sheet, they are dealt by \code{readxl()} with default
  arguments. \emph{sheetName} column will therefore be omitted.
}
\examples{
tidy_excel(readxl::readxl_example("type-me.xlsx"))
}
\seealso{
\itemize{ \item \link[dplyr]{mutate} \item \link[purrr]{map} }
}
