% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEDA.R
\name{waffleChart}
\alias{waffleChart}
\title{Draw a fast waffle chart}
\usage{
waffleChart(.tbl_df, .group, divisor = NULL, tiles = 100,
  title = NULL, size = 2, palette = "Set2", legend_pos = "right")
}
\arguments{
\item{.tbl_df}{A data frame}

\item{.group}{A \emph{factor} variable}

\item{divisor}{A non-zero positive number}

\item{tiles}{the default number of 100 for the number of tiles displayed on
plotting canvas if \code{divisor = NULL}}

\item{title}{A string}

\item{size}{A number.}

\item{palette}{A character vector or a string}

\item{legend_pos}{legend position}
}
\value{
A \strong{ggplot} object
}
\description{
\code{waffleChart()} is the flatten version of
  \code{pieChart()}. It works the best when a pie chart is less straight
  forward to reveal distribution of a factor variable.
}
\details{
A wrapper function inherits from
  \code{\link[waffle:waffle]{waffle()}}. It spares the data wrangling process
  for percentage computation and assemble for a named vector. The wrapper
  function also internally deals with conflict between the number of factor
  levels and the number of colors supplied to palette via
  \link[RColorBrewer]{RColorBrewer}.
}
\examples{
waffleChart(invest, age, divisor = 100)
}
